Let each CellRenderer choose if it must be added with expand to true or false.
It was a decission to hide CellRendering packagind into TreeViewColumns. This
leads to an easy API but, as a drawback, developers cannot choose whether to
add them with expand to true or false. In most cases, this is not a problem.
With text renderers it makes sense to use expand to true (specially on right
aligned text). Because that we had chosen to set expand to true as default.

However, with Pixbufs this is ugly, as white space appears around the image.
So we have finally chosen to let each CellRenderer type to choose what makes
more sense for it.