Switch to using cPickle
Performance optimization for the build system: it turns out that there
is a C implementation of the Pickle module in Python. Takes a build
with warm caches from ~2m15s to ~1m21s. Not bad.

You'd think they would just use this by default. Python is, after all,
a C program.