Code cleanup
Remove @Override annotations. They are totally unnecessary in an
environment where people use IDEs that provide visual indications.
Worse they are asymmetric: there is no point in using them on
overrides if there is no correponding one for implementing abstract
methods.

Fix copyright headers to refer to "program" instead of "library" in
the cases where it is a program (and GPLv2 licenced) and not part of
the library.