Internationalization test improvement.
Well, in fact this is not an improvement. It just happends that in many 
systems fr_CA locale is not installed, what causes tests to fail. I have 
changed it to en_CA, that seems to be installed by default in most 
systems. Anyway, this test is error-prone and we need to find a better 
way to test this...