Lock during builds
Build system improvement: problems arise when IDEs and humans are
overzealous and attempt to run multiple simultaneous builds - it
breaks the saving of the hashes pickle.

So, attempt to take an exclusive flock on a stamp file in tmp/ and
inhibit the build if that fails.

This is obviously frought with possiblity for problems, but as the
point is preventing multiple concurrent actions and not facilitating
multiple simultaneous access, we should be ok.