Fix getInsert() bug
Fix bug in TextBuffer's getInsert() method resulting from a cut and
paste error. 

Enhance test case generally exercising text insertion to evaluate
positioning of selection-bound and insert. 

Add coverage of TextBuffer's coverage of TextBuffer's selectRegion()
in order to be able to set these marks in order to test the
getSelectionBound() and getInsert() methods.