Change API for constructing TextTags
After creating every single TextTag you have to add it to a
TextTagTable before using it, with no reference to that TextTagTable,
in a TextBuffer. This is cumbersome.

Change the presented API so that when constructing TextTags you pass
the TextTagTable that you will be submitting it to, and have the
TextTag constructor automatically add it to the TextTagTable.