Get TextIters directly from TextMarks
Add a convenience method to TextMark allowing you easily get the
corresponding TextIter without having to mess around with the
TextBuffer explicitly, saving a few lines of code.

Document the caveats involved, especially as someone may use this as
an excuse to forget about the consequences of using TextIters
improperly.