Optional border Windows in TextViews
Preliminary coverage of the complex structure underlying the rendering
of TextView Widgets. Expose methods relating to the optional side
panels and the constants identifying them in TextWindowType. Discuss
extensively in that class and include a new snapshot attempting to
illustrate the optional Windows.

Methods allowing you to convert co-ordinates from buffer to
window origin and vice versa are also introduced. To follow the naming
schme used elsewhere in java-gnome for situations with multiple
out-parameters and to make the completion space sensible, the methods
have been named according to the pattern permutating 
convert{BufferToWindow,WindowToBuffer}Coords{X,Y}().