Fix bug in TextBuffer's setText()
The bindings layer had been incorrectly written to pass the number of
Unicode characters in a Java String as the length in bytes to the
third parameter of the function gtk_text_buffer_set_text(). This
resulted in incorrect length counts resulting in truncated strings and
test case failures.

The byte length of a UTF-8 encoded string is only available for a C
side jstring via (*env)->GetStringUTFLength(). We would have had to
write an override to call this manually, except that
gtk_text_buffer_set_text() provides for passing -1 as the length to
indicate that the string being passed in the second parameter is
already properly \0 terminated, which is acceptable for us as JNI
strings obtained via (*env)->GetStringUTFChars() are correctly NULL
terminated.

So hard code passing -1 as the length. Fixed.