Introduce the ability to connect signals after the default handler
In almost all instances signal handlers are conencted before the
exising, pre-registered, default handler. There are rare occasions,
however, where you need to offer the developer the possibility to
connect a singal after the default.

Introduce a boolean paramater after [corresponding to the gboolean
after in g_signal_connect_closure_by_id()] which propegates up through
org.gnome.glib.GObject's connectSignal() through to all the static
connect() methods generated on Widgets and other Object classes.

We have exposed this as a boolean at the translation layer, but this
is NOT to be exposed in the public API. Normal signal connection
methods will remain connect(); any time connecting after the default
is offered, the method is to be called connectAfter().