Prelminary treatment of invalidating areas and expose Rectangle
Cherry pick of from 'textview' branch:

Upgrade Rectangle so that they are allocated dynamically rather than
being a live reference to an underlying sub struct. (Recgangle was
previously only available via an EXPOSE_EVENT pointing into a
GdkEventExpose->area).

Expose invalidate() on org.gnome.gdk.Window, allowing us to use a
Rectangle to request that a given part of a Window be redrawn.