Enable getPixels() on a Pixbuf
Enable the ability to get an array of bytes representing the pixels in
a Pixbuf via its getPixels() method. This is not for writeback but
some purposes may require access to the raw pixel data.

This implementation is the result of *extensive* testing and
performance study; a number of alternative return types were
considered to get around the fact that Java interprets bytes as
signed, but ultimately the & 0xFF trick is sufficient to get around
that, so we ended up back at byte[].

The slightly adjusted nature of the return array is described in
detail in the JavaDoc and reinforced in the class comment.