Refactor array rebuilding to JNI code.
Slight performance improvement; the original implementation of
rebuilding the byte[] was in Java in the public API layer. This was
fine, but required a number of additional calls through JNI to
property getters which we were already making in our
gdk_pixbuf_get_pixels() override anyway. Under heavy load this takes
off about 5%.