Expose CONFIGURE_EVENT
Preliminary work to add CONFIGURE_EVENT to Window. It's here for the
same reasons that DELETE_EVENT is.

Add .defs data for GdkEventConfigure and expose EventConfigure as
needed by the signature of onConfigureEvent().

Did some preliminary testing with this and decided it was not what I
was looking for after all. That could mean it isn't documented
correctly.