Ensure Proxy lookup is thread safe
Anticipating greater concurrency with libraries that do not rely on
the GDK lock, properly synchronize access to the knownProxies table.

Obviously at the moment everything is subject to the GDK lock, but
that will change as we improve our Cairo coverage and introduce
GStreamer bindings.

Note that we're not using autoboxing; examining the byte code that
results via the disassembler and looking at the Open Java
implementation, Long.valueOf() does not gain us anything as it only
caches values of single byte size. Performance tests confirmed this,
although changing to something other than java.util.HashMap is
definitely in the cards.