Remove deprecated methods and incorrect virtuals from Style
The GtkStyle .defs data seems to have been the parking lot for a
considerable amount of GDK drawing functions. Remove all the
gdk_draw_* functions, and a few other deprecated methods as well.

By accident also noticed that many of these also were present as
virtuals. Certainly we don't need these as signals (there are only
two real ones), so they are removed as well.

Clearly this class predates much of the rest of GTK. Indeed, given
Cairo for drawing, its not clear how we can make use of the remaining
gdk_paint_* methods applying to a GdkStyle instance in our public API.