Correct a bug in the JNI location logic
The JNI location logic has code to strip the file name and /bin
directory name off of a $java executable, but there was a bug in the
regex, causing someone with a JAVA_HOME esque location including a
component named /bin to fail JNI header location.

Thanks to Patrick Ohnewein for the report and feedback that helped
find this unusual corner case.