Example of PDF drawing
Clone the drawing example to output to a PDFSurface. This was just an
initial experiment to see if anything actually came out. As is, it
not only shows up a number of deficiencies in Cairo's PDF rendering,
but also probably points out that the techniques used for drawing
onscreen are not really quite what one would do for drawing to
document. Figuring out these limitations will take quite a bit of
work.