Make the build process be silent when building from Eclipse
Add a silent variable to the build/faster script that can be toggled
on if the command line argument is "eclipse". This allows the IDE to
be configured to raise the console on changes to stdout without having
the build constantly doing so unnecessarily.

Fix a typo in the launcher that was preventing it from running in IDE
mode.