Add coverage of Viewport
Add preliminary coverage of the Viewport Bin.

Further work on Layout exposing its Ajustments. It seems to be a
combination of a Fixed that is scrolling capable (no Viewport
required) so I'm not sure what the difference is between that, and
just addWithViewport() a Fixed.

Miscallaneous improvements to ScrolledWindow, although they didn't
seem to work. Left in place not public with FIXMEs.