Possible workarounds for TextView's delayed line height calculations
Add a documentation section under the heading "Line Height
Calculations" to the top of TextView and warnings to getLineY() and
getLineRange(). Write a lengthy code snippet based on experiences with
a test case in another project.

This is obviously somewhat sketchy; if a better workaround appears, of
if this doesn't work for other people, we'll have to try something
else.