Restrict scope of builder
In Eclipse, I found that with a custom Builder as we have, the
external build was being triggered off quite frequently, often for no
good reason, and on occasion getting into build loops.

I configured a previously inactive feature, which is to specify a
working set of relevant resources. I'm not certain that this has been
done correctly. Please offer feedback if there are problems.