Comment out methods from PangoContext
pango_context_new should be only used to construct a new Pango
backend which relies on a special #define to be present in the C API
and pango_context_set_font_map is for internal use by Pango backends.
We will not be exposing this API in java-gnome so these methods have
been commented out.