Change name of internal static signal receiver methods
We did have the code generator cranking out names like handleClicked
and handleDeleteEvent(), but the prefix handle is too useful in
developer space to be used up this way. The rest of the code talks
about a "reveiver" layer, so change from that naming pattern to
receiveClicked() and recieveDeleteEvent(), etc.

This is all internal.