Documentation for TextBuffer
Write the first draft of the class documentation comment for
TextBuffer, describing how to insert text and the basic relationships
between TextBuffer, TextIter, and TextMark.

Resolve the problem of TextIter mutators being on TextView by exposing
them on TextIter. Pending further testing of course, but this is a
nice solution as it means all the methods relating to moving a
TextIter forward in a TextBuffer are available as forward<COMPLETE>.

Touch up the API stylesheet so that properties (which we have
stringently referred to with <var>...</var> tags) have a slightly
distinct colouring to differentiate them from normal emphasis in
italics.