Further documentation improvements, mostly relating to TextMark.
Expose two convenience constants on TextMark, LEFT and RIGHT, which
can be used when calling TextBuffer's createMark(). It was tempting to
create a strongly typed constant wrapping this, but, well, a) we
already have too many Gravity types going on, and b) {shrug} it's the
field name that really matters. Leave it as boolean.