Move to Proxy the responsability of calling release() on garbage collection.
Proxy defines a release() method each subclass should implement to free
underlying resources. Until now, the responsability to call this method on
finalize() was on Proxy direct subclasses (Object, Boxed...). I've moved it
to the natural place: Proxy.

This involves changes on Boxeds, that now must implement an abstract free()
method instead of release(), that, implemented in Boxed, only takes care
about handling the owner flag.

Note that this commit fixed a memory leak problem with Cairo bindings, but
it also shows a still unresolved problem on Entity subclasess.