Restart example code for TextView
Refactor the example demonstrating TextView usage to be a conversation
window from a little instant messaging program.

Change the constructor to be a series of method calls. Add an Entry
for the user to write messages that will be "sent". 

Place a TextView at the top of a window being the place where incoming
messages are displayed (and outgoing ones echoed). When text is being
put into the incoming message window, translate smileys from :) to a
happy face image. This is very cute.

Remove the rest of the code from the previous version of this example
that was no longer relevant.

Expose TextView's getBuffer().