Expose renderIcon()
This is a weird one. Although in the GtkWidget namespace [the native
function is gtk_widget_render_icon()] and despite the fact that the
first parameter is a GtkWidget*, this is really not a method on Widget.

So, expose it as a "function" on Gtk. This is quickly becoming the
home for all such general functions, and feels appropriate when coding
with it.