Expose the ability to display a Widget in a TextView
Clean up the TextChildAnchor and related methods TextBuffer's
createChildAnchor() and TextView's addChildAtAnchor().

Rename the latter to an add() override.

TextChildAnchor does not need to be public as we have done the create
and add steps together in one go in TextView.

I'd really prefer if this was insert() on TextBuffer instead of add()
on TextView. We'll see if we can figure that out.