Expose emission method for TreeView.RowActivated
It seems that passing NULL to gtk_tree_view_row_activated() is ok, so
adjust the .defs data to permit us to do so, quite frequently you want
to activate a row but don't much care which column.

The checkIter() method was bombing if null was passed. So, somewhat
redundently, test the TreeIter is not null before checking its validity.