Use full version in JNI library name
There is the possibility of a bug when someone has an older version of
java-gnome installed on their system (in /usr) but who is working on a
development version - the system installed shared library will get
picked up instead of the development one in tmp/ or wherever. Bad.

Since we have 100% tight coupling between the .so and the .jar, change
the library to load a .so with the full VERSION in its name rather
than just the APIVERSION. The native library at present will thus now
be libgtkjni-4.0.9.so and no longer libgtkjni-4.0.so.

This does not affect parallel installability; the library is slotted
to APIVERSION (ie 4.0) but does prevent the above bug. This should not
have any effect on downstream packages.