Enable parsing of characteristics with no value
While we were set up to handle lines like

  (deprecated "For this reason")

the DefsParser was choking on lines without a reason like

  (deprecated)

which was limiting as we want to stop generating code for methods we
don't need and don't particularly want to waste that much time
babbling on about why we don't need them.

Fixing this required significant surgery on the regexs in DefsParser,
adding a non-capturing group to the characteristics matcher and
removing the assumption that key/value will result in value not null.

For clarity, add a new characteristic line type for blocks we don't
want generated in the Java bindings:

  (unnecessary)

and rename DeprecatedException to UnnecessaryCodeException for better
readability.