Move native loading to Plumbing
Change the native library load from a System.loadLibrary() to a
System.load() with specified path. This path is formed from the value
of $libdir which is stored in a text file by configure. The key detail
is that this is NOT put into the .jar file located here in tmp/; it is
only added to the .jar when `make install` is done. This way the
completed build can be used in-place in a checkout of java-gnome
without having to install it.

Move translation layer implementations relating to the Gtk entry point
class to new class GtkMain. The primary purpose of this is to be a
plumbing subclass so we force class initialization right up to
org.freedesktop.bindings.Plumbing.

Move the Version constants file to org.freedesktop.bindings. Wrap the
constants in getters (which should help the bugs relating to the
propegation of those values changing).