Work out location of .jar in fallback case and load .so from there
Go to quite some trouble to work out what the location of the .so will
be with an "in-place" build. Since they can be used from an arbitrary
external location, we need to work out the URL of the .jar file that
was referenced in the first place and use that to find the directory
to use as libdir. This involves some pretty voodoo code that I don't
entirely trust.

Eclipse no longer needs the native library location specified. Hooray!