Expose TreeView's getPathAtPos()
Add coverage for TreeView's getPathAtPos() getting a TreePath from x,y
co-ordinates and a corresponding getColumnAtPos() for the
TreeViewColumn out-parameter of the underlying function. Correct the
.defs data based on what is written in the GTK doc.

Document extensively the case where you are trying to create a right
click context menu, and the need to do a workaround to ensure that the
row being clicked is properly selected.