Update EventButton's getX() and getY() to return doubles instead of ints
EventButton's getX() and getY() methods are now set to return doubles. The GDK
C Library uses double type property values for a GdkEventButton's X and Y. The
rationale is that, for applications and interfaces where subpixel accuracy
is required, such as on touchscreens and tablet PCs, it is better to let the
application decide if it should round off/truncate the X and Y values.

Also updated was an example in TreeView's Javadoc.