Possible fix for the problem of the test suite showing a Window
The main loop cycling that happens in the ValidateScreenAndDisplays
TestCase causes a Window to present briefly, however this isn't the
cause. The offending code seems to be in ValidateProperties where we
are testing Allocations and need the Widget to have been through the
size-request / size-allocate cycle.

Instead of show() on the top level, trying realize() [still internal
as of this writing] seems to make the problem go away. Since realize()
isn't actually exposed yet (and may not ever be), this fix seems a bit
weak.