Access the extents of a Pango Layout and render single lines
Add an overload to showLayout() on the Cairo Context allowing you to
render an individual LayoutLine.

Expose methods providing access to the utility methods which get at
the extents of laid-out text on both Layout and LayoutLine.

Add wrappers to Rectangle representing ascent and descent
which are perhaps more convenient.