Add three holes to make this three-hole-punch paper
Draw three circles and fill them with white, blotting out the blue
rules and making them look a bit like holes punched out of the piece
of paper. 

Rearrange the Layout drawing to the end of the program so that the
text lies above the blue lines.

Fix up the documentation of Context's fillPreserve() and expose
strokePreserve() while we're at it.