Fix crash in org.freedesktop.cairo.Plumbing.createSurface()
In our JNI code which differentiates between Cairo types and creates
the approriate proxy object, we were caching the results of the jclass
lookups. This was fine, except that the return value from FindClass()
is a local reference, not a global one. Oops.

So add calls to NewGlobalRef() around the result of the FindClass()
lookups in both createSurface() and createPattern().