Illustration of arc()
Create a snapshot for Context showing its arc() method and the
(counter-intuitive) fact that increasing angles are clockwise, not
counter clockwise.

Add coverage of the relative moving and line drawing methods, along
with the ability to get the co-ordinates of the current point.

Add NO_CURRENT_POINT to the Status error constants.

Enable the code generator to handle double out-parameters; that was
missing along the way somehow.