Add coverage of Cairo Matrix and transforms
Expose Cairo's cairo_matrix_t as Matrix. Implement an override to
allocate and free these structures. The allocator calls the
function to initialize to the identity transformation.

Add a Snapshot for each of scale(), rotate(), and translate(). Do some
refactoring of the parent classes so that we have some better code
reuse.