Fix memory management of LayoutLine instances
Properly increase the reference count of a PangoLayoutLine when we
create our Proxy for it. Failing to do so was resulting in an eventual
VM crash originaing somewhere within Pango, but the culprit was fairly
clearly our not ref'ing the PangoLayoutLine that we keep ahold of.

Note that we had to do away with the return value to avoid generating
code which would result in an infinite loop.