Refine Pixbuf from data constructor
Add necessary memory management check to JNI code. This is a
constructor; the Ref needs to be taken.

Implement the usual private static wrapper to do Exception trapping
and handling in a constructor.

Add a test fixture to the ValidateImageHandling test case class
ensuring that various error states are indeed trapped and result in
Exceptions.