Add strict option to top level configure
Ubuntu's J version is shipping GTK 2.15 which has an enormous amount
of deprecations. Our previous pattern was to depend on equal to the
current stable release, but with them shipping unstable this causes
problems for us as we were automatically enabling flags like
GTK_DISABLE_DEPRECATED for anyone with a Bazaar checkout.

Hackers can now specify:

$ ./configure strict

to disable the deprecated code in the GNOME libraries for working on
removing any such reliances from our code.