Expose a Notify signal
Add coverage of TextBuffer.NotifyCursorPosition, being a signal
emitted when the 'cursor-position' property of a TextBuffer changes.

This is a bit of a prototype for how notify signals could be done in
java-gnome.

Engineering changes to the signal marshalling JNI code in
bindings_java_signal.c so that parameters of type GParamSpec can be
ignored. As it stands right now we don't actually need to model their
values on the Java side as developers can use property getters.

Having exposed TextBuffer.NotifyCursorPosition, we can recommend it in
preference to TextBuffer.MarkSet.