Pass a void* instead of JNIEnv* to GetEnv()
Warnings were being emitted from src/jni/bindings_java_util.c when
compiling under -O2. Thanks to Serkan Kaba for having raised the issue.

The problem was apparently due to our passing a different struct to
the one expected by the function prototype. Mark Wielaard suggested
the cause possibly using a union to hint the compiler and avoid the
warning.

This patch is also a possible fix for the problem, instead attempting
to pass a pointer of the expected type and casting it afterwards.