Optimize setting Attribute indexes
Replace public setters for start_index and end_index with a utility
method called setIndexes(). Create an override so that C side we can
use the already converted to UTF-8 text of a PangoLayout to calculate
the offsets.

Rename our public API class representing PangoAttrList to
AttributeList, better matching Attribute since we don't have Attr
exposed as a prefix anywhere.