Create TextBuilder to wrap use of Attributes
Pango Attributes have a problem: you need to set the start and end
offsets that they apply to. This is a nightmare for us because in Java
we have no knowledge of the UTF-8 offset a given character position
corresponds to (and no desire to expose it whatsoever; type safety!)

Create new class TextBuilder which wraps a StringBuilder and an
AttributeList, allowing you to build up the consecutive
(String,Attribute) pairs to be rendered by a Layout.