Marshall gpointers as NULL
In the event we get a signal parameter of type G_TYPE_POINTER (which
is how the internal type GtkNotebookPage* is registered in the public
signal 'switch-page') marshall it across the boundary as a type 'J'
and thence value NULL -> (jlong) 0.

On the Java side you'll have to code something to remove it from the
parameter list, but at least the program won't fail to run.