Change main() to take GDK lock on JNI side
Atypically we do the necessary operations to take and release the GDK
lock here on the JNI side; everywhere else in the library we use a
Java side synchronized block.

This works around a strange behaviour in Eclipse where execution was
suspending at synchronized block and not where the Exception was being
thrown, and hopefully results in a better debugging experience.