Rename getLocation()
Expose gtk_text_view_get_iter_location() as TextView's getLocation().

We decided not to refactor the name of getVisibleRectangle() although
this method is complementary and both return Rectangle objects. It was
tempting, but probably best to leave getVisibleRectangle() alone.
Meanwhile, this method could not be called getIterLocation() as all
the other getIter() methdos return a TextIter.