Fix memory management of Pango Attributes
It turns out that when you add a Pango Attribute to a AttributeList via
its insert() method, responsibility to free that Attribute is assumed
by the list. This is complicated to deal with since there is no
reference count mechanism on such structs. So we have to do it
ourselves.

Add a unit test (which also happens to includea main() entry point)
running up some example text marked up with Attributes, and then make
sure that the VM doesn't crash on a run of the garbage collector.