Dramatic improvement to Attribute handling
Change the way Attribute offsets are set. Hold the offset values on
the Java side and only convert them to UTF-8 when the call is made to
Layout's setAttributes(). Actually do the PangoAttrList insertion at
that time.

This means that Attributes can be created while accumulating
characters in a StringBuffer without needing to have set the entire
text body into the Layout.

This, along with the crash fixes, is important enough to drive a new
java-gnome release.