Expose Application.MessageReceived signal
Write the machinery to hoop up Application.MessageReceived signals,
including the necessary wrapper class to turn command ordinals into
Command constants.

Write up a trivial example demonstrating unique functionality in
action. This is probably not complete in that it had some strange
interactions when I tried to do GUI work with it.