Correct character retrieval
ABI BREAK

The code generator had the wrong mapping for gunichar. It represents
a code point needs to be an int not a char.

This primarily affects TextIter's getChar(). Correct the signature of
the public API of this method to return int. Update unit tests. Mark
as now being @since 4.0.13. This is an ABI breakage; any code using
this method will need to be recompiled.