Prototype stopping a signal emission
Expose two new methods, stopInsertText() and stopDeleteRange(), which
are manually backed by calls to g_signal_stop_emission(). This
prototypes a public API for exposing the ability to stop a signal from
further propegation.

For the case of TextBuffer.InsertText, you may want to halt the
insertion that results from a user typing into a TextView. This allows
you to, in effect, block the default handler which carries out the
actual insertion of the typed character into the TextView.