Handle GdkKeyval as if it were typedef guint
GdkKeyvals are something like an enum, but are passed as guint
throughout the C API; there is no type declared GdkKeyval. But if we
pretend there is, then we can have the code generator do all the
ordinal extracting for us.

Create a new Thing subclass called TypedefEnumThing to allow us to
register this type.

Introduce .defs data for GtkTest, holding a define block for
gtk_test_widget_send_key(), which takes such a GdkKeyval as an
argument.