Run tests in virtual X server
Import code from Quill and Parchment to set up an instance of the Xvfb
virtual X server to run the java-gnome test suite against.

The entire reason we did this was so that we could test InputMethods,
which need keystrokes sent to Widgets in a shown Window. However, for
some reason, while normal keystrokes work, sending Compose via
gtk_test_widget_send_key() to a Window in an Xvfb fails.

Add a C program to demonstrate this.