Minor correction to C code demonstrating Compose problem in Xvfb
When using gtk_im_context_filter_keypress() you also need to hookup
the callback to "key-release-event"; we're doing that in our Java code
but left it out of the exploration in C when we were demonstrating the
bug.

This doesn't change anything; the problem is still extant when this
program is run in a virtual X server.