Avoid test popping up a web browser
Can't call LinkButton's inherited emitClicked() because that results
in a web broswer popping up. Never do. Move that to a skipped test
fixture.

It's be nice if we could trap that into the virtual X server somehow.
Maybe this signifies our test environment isn't quite as isolated as
we might like.