Expose LayoutLine's start and length properties
Write overrides so that we can get to the start_index and length
fields of a PangoLayoutLine struct and convert them to character
offsets (since we don't work in bytes in our public API).

Copy the basic structure of ValidatePangoAttributes to create
ValidatePangoLayoutLines which tests that our line breaking and thence
offset conversions are behaving.