Deprecate no-arg Box pack methods
The true, true "default" we had in packStart() and packEnd() were not
helping things and leading people to consistently the wrong thing.
They have no equivalent in GTK, and this was causing confusion and
bugs in the examples. Enough. People need to know about the expand
property.

Deprecate packStart(Widget) in favour of the full packStart(Widget,
boolean, boolean, int) and likewise packEnd(Widget) for
packEnd(Widget, boolean, boolean, int).