Fix image data cleanup
Immediately releasing data array was removing the data from Cairo's
reach. So use a transfer struct to hold a reference to the
jbyteArray and the jbyte* for later cleanup.

Not much of a test case, but doing System.gc() was triggering a crash
here until we got this right, so the code path is being used.