Ensure GdkWindow properly ref'd
Our implementation code for Widget's getWindow() was not taking a Ref,
with the result that when our memory management code added a ToggleRef
and dropped the Ref we were one down, leading to premature freeing of
the GdkWindow.

Could have repaired our override with a call to
bindings_java_memory_cleanup(), but now that there is a strong
accessor gtk_widget_get_window() we can just generate code for that
instead.