Emit alert in logging handler
We have happily been throwing our FatalError Exception when someone in
GTK etc logs a CRITICAL or WARNING message. However, the Java VM messes
with our good intentions because Exceptions which occur during finalizing are
swollowed. Grr.

Since these messages are often the root cause problem, it's important
to hear about them. So we output an alert to stderr before throwing.
Not thrilled about that.