Add coverage of GTK Object's destroy()
Unexpectely, gtk_object_destroy() does NOT have immediate memory
management implications. It really does do just what it's docs say -
asks things holding a Ref to this GtkObject to relinquish it. Fair
enough, and it is for numerous reasons preferable to
gtk_container_remove() if you're done with a Widget. Since it's not
"close" in the usual sense, why not. Add Object.Destroy signal as
well.