Force static initalizers and/or library initialization code in setUp()
Move initializers, being run once in an initial test case, to a
setUp() block. I'm a bit vague as to why this is necessary; it would
seem that JUnit or Java 7 is being stricter about its classloading, so
we need to force the static initializers to run every test. This is
somewhat surprising, but seems a workaround. It's more "correct" JUnit
in any event.