Allow IconView to be added to ScrolledWindow using add()
IconView has scrolling support, so it shouldn't have to be added using
addWithViewport(). When you use addWithViewport(), the IconView does
not correctly re-flow icons as the ScrolledWindow is resized; it
re-flows when the window is widened, but adds a horizontal scrollbar
when the window is narrowed.

A demo program illustrates this issue: http://pastebin.com/HajMj2uh